+++ backend/packages/eks/CHANGELOG.md
@@ -1,5 +1,31 @@
 # @cat-factory/eks
 
+## 0.1.275
+
+### Patch Changes
+
+- Updated dependencies [11f9efa]
+  - @cat-factory/contracts@0.278.0
+  - @cat-factory/integrations@0.152.3
+  - @cat-factory/kernel@0.275.2
+
+## 0.1.274
+
+### Patch Changes
+
+- Updated dependencies [c44e9d7]
+  - @cat-factory/contracts@0.277.0
+  - @cat-factory/integrations@0.152.2
+  - @cat-factory/kernel@0.275.1
+
+## 0.1.273
+
+### Patch Changes
+
+- Updated dependencies [dfa4a8e]
+  - @cat-factory/kernel@0.275.0
+  - @cat-factory/integrations@0.152.1
+
 ## 0.1.272
 
 ### Patch Changes